add model factory
